third_party/blink/web_tests/fast/table/table-with-borderattr-null-expected.txt

There should be a black box below.


PASS getComputedStyle(document.querySelector("table")).borderTopWidth is "1px"
PASS successfullyParsed is true

TEST COMPLETE
